refactor: BController로 분리
