+++ src/main/java/com/ll/medium/domain/post/post/controller/PostController.java
@@ -1,6 +1,5 @@
 package com.ll.medium.domain.post.post.controller;
 
-import com.ll.medium.domain.member.member.entity.Member;
 import com.ll.medium.domain.member.member.service.MemberService;
 import com.ll.medium.domain.post.post.entity.Post;
 import com.ll.medium.domain.post.post.service.PostService;
@@ -16,6 +15,7 @@ import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 
 @Controller
+@RequestMapping("/post")
 @RequiredArgsConstructor
 public class PostController {
     private final PostService postService;
@@ -23,35 +23,14 @@ public class PostController {
     private final Rq rq;
 
     // 글 상세 페이지 조회
-    @GetMapping("/post/{id}")
+    @GetMapping("/{id}")
     public String showPost(@PathVariable long id){
         rq.setAttribute("post", postService.findById(id).get());
         return "domain/post/post/detail";
     }
 
-    // 사용자가 작성한 글 목록 조회
-    @GetMapping("/b/{username}")
-    public String showUserPostList(@PathVariable String username){
-        // 해당하는 회원이 없는 경우 예외 처리
-        if(!memberService.findByUsername(username).isPresent()) throw new RuntimeException("해당하는 회원이 없습니다.");
-        rq.setAttribute("posts", postService.findByAuthorUsernameAndIsPublishedOrderByIdDesc(username, true));
-        return "domain/post/post/userPostList";
-    }
-
-    // 사용자가 작성한 특정 글 상세 페이지 조회
-    @GetMapping("/b/{username}/{id}")
-    public String showUserPostDetail(@PathVariable String username, @PathVariable long id){
-        Post post = postService.findById(id).get();
-        // 글 접근 권한이 없는 경우 예외 처리
-        Member member = memberService.findByUsername(username).get();
-        if(!postService.canAccessPost(post, member, id)) throw new RuntimeException("해당하는 글이 없거나 비공개 된 글입니다..");
-        rq.setAttribute("post", post);
-
-        return "domain/post/post/detail";
-    }
-
     // 전체 공개된 글 목록 조회
-    @GetMapping("/post/list")
+    @GetMapping("/list")
     public String showList(){
         rq.setAttribute("posts", postService.findByIsPublishedOrderByIdDesc(true));
         return "domain/post/post/list";
@@ -59,7 +38,7 @@ public class PostController {
 
     // 현재 사용자가 작성한 글 목록 조회
     @PreAuthorize("isAuthenticated()")
-    @GetMapping("/post/myList")
+    @GetMapping("/myList")
     public String showMyList(){
         rq.setAttribute("posts", postService.findByAuthorIdOrderByIdDesc(rq.getMember().getId()));
         return "domain/post/post/myList";
@@ -67,7 +46,7 @@ public class PostController {
 
     // 글 작성 페이지 조회
     @PreAuthorize("isAuthenticated()")
-    @GetMapping("/post/write")
+    @GetMapping("/write")
     public String showWrite(){
         return "domain/post/post/write";
     }
@@ -85,7 +64,7 @@ public class PostController {
 
     //글 수정 페이지 조회
     @PreAuthorize("isAuthenticated()")
-    @PostMapping("/post/write")
+    @PostMapping("/write")
     public String write(@Valid WriteForm writeForm){
         RsData<Post> writeRs = postService.write(rq.getMember(), writeForm.getTitle(), writeForm.getBody(), writeForm.getIsPublished());
 
@@ -94,7 +73,7 @@ public class PostController {
 
     // 굴 수정 처리
     @PreAuthorize("isAuthenticated()")
-    @GetMapping("/post/{id}/modify")
+    @GetMapping("/{id}/modify")
     public String showModify(@PathVariable long id){
         Post post = postService.findById(id).get();
         // 글 수정 권한이 없는 경우 예외 처리
@@ -116,7 +95,7 @@ public class PostController {
     }
 
     @PreAuthorize("isAuthenticated()")
-    @PutMapping("/post/{id}/modify")
+    @PutMapping("/{id}/modify")
     public String modify(@PathVariable long id, @Valid ModifyForm modifyForm){
         Post post = postService.findById(id).get();
         // 글 수정 권한이 없는 경우 예외 처리
@@ -129,7 +108,7 @@ public class PostController {
 
     // 글 삭제 처리
     @PreAuthorize("isAuthenticated()")
-    @DeleteMapping("/post/{id}/delete")
+    @DeleteMapping("/{id}/delete")
     public String delete(@PathVariable long id) {
         Post post = postService.findById(id).get();
